Add files via upload
Was running into trouble with code and tried to fix it last minute.
Couldn't fix it, but I prefer this code better than the last one.
If editting it after the due date is considered a zero, I apologize. I'm willing to take the 0 on this assignment if so. Please let me know in the comments when grading! Thank you